+++ matrix-sdk-android-rx/src/main/java/im/vector/matrix/rx/RxRoom.kt
@@ -19,6 +19,7 @@ package im.vector.matrix.rx
 import android.net.Uri
 import im.vector.matrix.android.api.query.QueryStringValue
 import im.vector.matrix.android.api.session.events.model.Event
+import im.vector.matrix.android.api.session.identity.ThreePid
 import im.vector.matrix.android.api.session.room.Room
 import im.vector.matrix.android.api.session.room.members.RoomMemberQueryParams
 import im.vector.matrix.android.api.session.room.model.EventAnnotationsSummary
@@ -102,6 +103,10 @@ class RxRoom(private val room: Room) {
 
     fun invite(userId: String, reason: String? = null): Completable = completableBuilder<Unit> {
         room.invite(userId, reason, it)
+    }
+
+    fun invite3pid(threePid: ThreePid): Completable = completableBuilder<Unit> {
+        room.invite3pid(threePid, it)
     }
 
     fun updateTopic(topic: String): Completable = completableBuilder<Unit> {