+++ vector/src/main/java/im/vector/riotx/features/createdirect/CreateDirectRoomViewModel.kt
@@ -22,8 +22,9 @@ import com.airbnb.mvrx.ViewModelContext
 import com.squareup.inject.assisted.Assisted
 import com.squareup.inject.assisted.AssistedInject
 import im.vector.matrix.android.api.session.Session
-import im.vector.matrix.android.api.session.room.model.create.CreateRoomParams
+import im.vector.matrix.android.api.session.room.model.create.CreateRoomParamsBuilder
 import im.vector.matrix.rx.rx
+import im.vector.riotx.core.extensions.exhaustive
 import im.vector.riotx.core.platform.VectorViewModel
 import im.vector.riotx.features.userdirectory.PendingInvitee
 
@@ -53,11 +54,17 @@ class CreateDirectRoomViewModel @AssistedInject constructor(@Assisted
     }
 
     private fun createRoomAndInviteSelectedUsers(selectedUsers: Set<PendingInvitee>) {
-        val roomParams = CreateRoomParams(
-                invitedUserIds = selectedUsers.map { it.userId }
-        )
-                .setDirectMessage()
-                .enableEncryptionIfInvitedUsersSupportIt()
+        val roomParams = CreateRoomParamsBuilder()
+                .apply {
+                    selectedUsers.forEach {
+                        when (it) {
+                            is PendingInvitee.UserPendingInvitee     -> invitedUserIds.add(it.user.userId)
+                            is PendingInvitee.ThreePidPendingInvitee -> invite3pids.add(it.threePid)
+                        }.exhaustive
+                    }
+                    setDirectMessage()
+                    enableEncryptionIfInvitedUsersSupportIt = true
+                }
 
         session.rx()
                 .createRoom(roomParams)