+++ vector/src/main/java/im/vector/riotx/features/userdirectory/UserDirectoryFragment.kt
@@ -21,6 +21,7 @@ import android.view.View
 import com.airbnb.mvrx.activityViewModel
 import com.airbnb.mvrx.withState
 import com.jakewharton.rxbinding3.widget.textChanges
+import im.vector.matrix.android.api.session.user.model.User
 import im.vector.riotx.R
 import im.vector.riotx.core.extensions.cleanup
 import im.vector.riotx.core.extensions.configureWith
@@ -81,9 +82,9 @@ class UserDirectoryFragment @Inject constructor(
         directRoomController.setData(it)
     }
 
-    override fun onItemClick(pendingInvitee: PendingInvitee) {
+    override fun onItemClick(user: User) {
         view?.hideKeyboard()
-        viewModel.handle(UserDirectoryAction.SelectPendingInvitee(pendingInvitee))
+        viewModel.handle(UserDirectoryAction.SelectPendingInvitee(PendingInvitee.UserPendingInvitee(user)))
         sharedActionViewModel.post(UserDirectorySharedAction.GoBack)
     }